requirements.txt
autopep8==2.0.0
numpy==1.23.4
Pillow==9.3.0
pycodestyle==2.9.1
tk==0.1.0
tomli==2.0.1
